build: Change xmlrpc-c to super-stable branch
The purpose of this project is for seed box platforms. As a result, we should prioritize stability over performance.

There shouldn't be any significant impact for this dependency for rTorrent software performance. We have a lot of gain in terms of stability on the super stable branch.

We need to minimize the amount version bumps to ensure the xmlrpc-c developer has an adequate opportunity to test changes before they get pushed to upstream. Unstable changes is a very big problem for this software. The impact should be mitigated all costs to keep platforms stable.