rTorrent: Simplify curl dual stack IPv4 / IPv6 handling
From anthonyryan1:
https://github.com/rakshasa/rtorrent/commit/ae1d46252c4da8149009fc4c40ca899fe4dcb8dc

curl implements the RFC 6555 "Happy Eyeballs" algorithm, meaning a
manual fallback to IPv6 is uncessary and arguably detrimental.

If we leave curl to it's own devices with the request, it will attempt
to connect on both IPv4 and IPv6, then use whichever connection is
going to be more reliable.