rTorrent: Add tracker delay scrape feature
Resolves https://github.com/rakshasa/rtorrent/issues/1044

With thousands of public tracker torrents (specially UDP), rTorrent will peg a single core at 100% CPU usage until the thread times out and the software terminates.

With hundreds of public tracker torrents (specially UDP), rTorrent will take 5 minutes to launch wwhile it scrapes all the trackers.

To resolve this problem, I have implemented a tracker delay scrape feature. It will start scraping trackers after the rTorrent software has been initialized.

The benefit of this change is instantaneous startup times and immediate access to the torrent client through ruTorrent. There is no more delays or software crashes.

The only downside of this change is it will take a couple more minutes to announce to all the trackers with thousands of torrents. This is tolerable given the benefits it provides.

The feature has been enabled by default for the objective of a seed box platform. It improves the stability of the torrent client.

A configuration option is also added, should the user which to toggle the feature. It can be disabled by specifying the following value in the .rtorrent.rc file.

trackers.delay_scrape.set = no