libtorrent: Add scanf cache vector (#5)
This commit creates a new udp tracker info vector to parse the hostname and port for UDP trackers.

It avoids calling sscanf tens of thousands of times, which may result in performance issues. Worse case scenario, it can also crash the software.

It's often the case where multiple torrents will have the same UDP trackers. If this is the case, the hostname and port will be recalled from the vector to avoid repetitive lookups. This improves the performance when initializing torrents.

The hostname and port data is then passed to the UDP tracker class on initialization.