libtorrent: Simplify updating queue entries (#6)
From kannibalox:
https://github.com/rakshasa/libtorrent/pull/248

The erase/insert pattern can be replaced by a single method that updates an existing entry's timer, or inserting if it doesn't exist.

This primary goal was to reduce CPU usage when handling incoming connections on large instances, but since the optimization was pretty generic it's been applied to many places.

Notes from stickz: This patch leads to higher throughput when seeding torrents. It greatly reduces the CPU usage of the rTorrent software. It is 100% stable and there are no downsides to using the patch. Tracker updates are also performed faster.