libtorrent: Increase download delegator block size
We can greatly improve the download throughput on multi-gigabit connections by increasing the block size on the download delegator. There is a significant bottleneck associated with processing too many blocks at once. This also reduces CPU usage.

Please ensure to set TCP socket sizes to 32MB in order to properly utilize this change. It will push more data to the network stack.

Estimated results are a 30% reduction is overall CPU usage per gigabit and a 50% increase in download throughput.

We should not set this value any higher. Instead we should optimize the download delegator in other ways. It will start sending too much information to the network stack.

The hashing thread is able to handle well over 10 gigabits. However, there is still a source code bottleneck with the download delegator. We should be able to make generic optimizations.