rTorrent: optimize loading session torrents
from jesec
https://github.com/jesec/rtorrent/pull/37

main: batch checking after the whole session loaded

"receive_hashing_changed" is an O(n) operation, where n is the
number of entities in the download list. It is not efficient when
the function is triggered every time a torrent has been inserted
by "load_session_torrents", making the whole operation O(n^2).

Thus, delay checking until the whole session has been loaded.
view: add a fast path for newly added element

filter_download is often called after a new element is added.

Since new elements are always inserted to the back, it is more
efficient to check the last element first, so we have O(1) instead
of O(n) in this common case.
view: insert directly if sorting is not required

O(n) -> O(1)
main: only set view sorting and periodic resorting when not in daemon mode

Sorting is O(n*log(n)) and very expensive. RPC user (such as Flood
or ruTorrent) does not rely on rTorrent for sorted list.
download_factory: don't check for "removed-right-after-add" when loading session

As no command is attached when loading from session, it is
impossible for this edge case to occur.

"find" on the download list, a "std::list" linked list, is
an O(n) operation. With it, the session loading is O(n^2),
where n is the number of entities in the session directory.

As such, remove this unnecessary check to speed things up.
main: bypass queuing in "load_session_torrents"

It is more expensive to queue all the items and then dequeue and
process them one by one, due to O(n*log(n)) time complexity of
"priority_queue", where n is the number of tasks already in the
queue, and memory reallocation.

We do not need task queuing in the synchronous loading of session.
After all, we called "priority_queue_perform" right after queuing.

As such, call "f->set_immediate(true)", so at any time the queue
has at most 1 task, which speeds up session loading.