libtorrent: Add fallocate all feature
from jesec:
https://github.com/jesec/libtorrent/commit/46b32949f35c32c518a1b634d6ec9cab0fe2580d

In many cases, it is preferable to preallocate space for the whole
torrent at once, to reduce fragmentation of contents inside the
torrent.

It allows more efficient filesystem organization, where files that
are likely to be read together are packed in contiguous area of the
disk, improving the performance and I/O patterns of future read
operations.