libtorrent: Implement two hash queue threads (#28)
We can evenly divide the downloads between two different HashQueue objects. This allows us to create anther disk thread to schedule the CPU load more evenly between multiple processor cores. A common scenario is for 30% of the CPU usage to be allocated for hashing. This pull request would split that load between two threads at 15% when multiple downloads are active.