libtorrent: Fix crash with null TrackerUdp object
This commit sends a pointer for TrackerUdp object (which holds the callback for the UDNS query) to libudns then back again. On arrival back to libtorrent, it checks if the pointer to the TrackerUdp object is valid before calling the UDNS callback. This avoids a crash by not trying to invoke a function which is no longer valid.